+++ src/modules/posts/infra/typeorm/entities/Post.ts
@@ -42,7 +42,7 @@ class Post {
   @Column({ default: false })
   notified: boolean;
 
-  @Column({ type: 'integer', array: true, default: [] })
+  @Column({ type: 'bigint', array: true, default: [] })
   notified_to: number[];
 
   @Column({ default: false })
